+++ ceammc/ext/src/dyn/dyn_gate.cpp
@@ -1,37 +1,13 @@
 #include "dyn_gate.h"
-#include "ceammc_crc32.h"
 #include "ceammc_factory.h"
+#include "dyn_gate_base.h"
 using namespace ceammc;
 
-class DynGate : public faust_dyn_gate_tilde {
-    UIProperty* threshold_;
-
+class DynGate : public DynGateT<faust_dyn_gate_tilde> {
 public:
     DynGate(const PdArgs& args)
-        : faust_dyn_gate_tilde(args)
+        : DynGateT<faust_dyn_gate_tilde>(args)
     {
-        CEAMMC_DEFINE_PROP_VAR(threshold)
-        CEAMMC_DEFINE_PROP_VAR(attack)
-        CEAMMC_DEFINE_PROP_VAR(hold)
-        CEAMMC_DEFINE_PROP_VAR(release)
-
-        threshold_ = findUIProperty(sym_prop_threshold);
-        checkUIProperties({ threshold_ });
-
-        bindPositionalArgsToProps({
-            sym_prop_threshold,
-            sym_prop_attack,
-            sym_prop_hold,
-            sym_prop_release,
-        });
-
-        createInlet();
-    }
-
-    void onInlet(size_t n, const AtomListView& lv)
-    {
-        if (threshold_)
-            threshold_->setList(lv);
     }
 };