+++ ceammc/ext/src/dyn/dyn_gate2.cpp
@@ -1,24 +1,24 @@
 #include "dyn_gate2.h"
 #include "ceammc_factory.h"
+#include "dyn_gate_base.h"
 using namespace ceammc;
 
-class DynGate2 : public faust_dyn_gate2_tilde {
+class DynGate2 : public DynGateT<faust_dyn_gate2_tilde> {
 public:
     DynGate2(const PdArgs& args)
-        : faust_dyn_gate2_tilde(args)
+        : DynGateT<faust_dyn_gate2_tilde>(args)
     {
-        bindPositionalArgsToProps({ gensym("@threshold"),
-            gensym("@attack"),
-            gensym("@hold"),
-            gensym("@release") });
     }
 };
 
 void setup_dyn_gate2_tilde()
 {
     SoundExternalFactory<DynGate2> obj("dyn.gate2~");
+    obj.setXletsInfo(
+        { "signal: left input", "signal: right input", "float: threshold (dbfs)" },
+        { "signal: left output", "signal: right output" });
 
     obj.setDescription("stereo signal gate");
     obj.setCategory("dyn");
-    obj.setKeywords({"gate"});
+    obj.setKeywords({ "gate" });
 }